+++ CHANGELOG.md
@@ -22,6 +22,8 @@ CHANGELOG
     a shim based on `$.ajax()` (requires jQuery 3.0 or higher).
   - Please check the documentation for the new options that can be passed.
 - Renamed the option `suppressMouseWheelSelector` to just `suppressWheelSelector`.
+- Improved submenu positioning by automatically opening them on the left-hand side if there's
+  insufficient space on the right side.
 - Fix #107: Remove the dropdown after timeout to fix "hover" behavior.
 - Fix #136: Update original `<select>` element on "change" instead of "selectivity-selected".